✨ feat: Refactor Artifacts component structure and add mobile/desktop variants for improved UI
